Add styling to the checkout-address form, Rename address component to checkout-address page
