+++ app/cart/page.jsx
@@ -5,7 +5,6 @@ import { CartContext } from "../components/CartContext"
 import axios from "axios"
 import Link from "next/link"
 import Header from "../components/Header"
-import CheckoutButton from "../components/CheckoutButton"
 
 const Cart = () => {
   const { cartProducts, addItemToCart, decrementItemInCart, removeItemFromCart } = useContext(CartContext) //The cartProducts will have: productId and quantity
@@ -118,7 +117,11 @@ const Cart = () => {
                         (Taxes and Delivery charges will be added in the checkout page)
                       </div>
                       <div className="text-center">
-                        <CheckoutButton />
+                        <button className="text-white text-lg py-2 px-4 rounded-md focus:outline-none">
+                          <Link href={'/checkout-address'}>
+                            Checkout
+                          </Link>
+                        </button>
                       </div>
                     </div>
                     )}
